requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
sentence-transformers
faiss-cpu
requests